Debugging the CI fail
